LogiRoute/requirements.txt
mysql-connector-python
openai
pandas
datasets
ragas
langchain-openai
langchain-huggingface
langchain-core
requests
marker-pdf
python-docx
python-dotenv
